# Cache char ASCII equivalence to speed-up transliteration