--- a/tox.ini
+++ b/tox.ini
@@ -1,0 +1,11 @@
+[tox]
+envlist = {py27,py34}
+skipsdist = True
+
+[testenv]
+basepython =
+    py27: python2.7
+    py34: python3.4
+
+commands =
+    python setup.py test