--- a/unidecode/__init__.py
+++ b/unidecode/__init__.py
@@ -13,7 +13,9 @@ In Python 3, a standard string object will be returned. If you need bytes, use:
 >>> unidecode("Κνωσός").encode("ascii")
 b'Knosos'
 """
-Cache = {}
+Cache_sections = {}
+Cache_chars = {}
+
 
 def unidecode(string):
     """Transliterate an Unicode object into an ASCII string
@@ -25,30 +27,35 @@ def unidecode(string):
     retval = []
 
     for char in string:
-        codepoint = ord(char)
-
-        if codepoint < 0x80: # Basic ASCII
-            retval.append(char)
-            continue
-        
-        if codepoint > 0xeffff:
-            continue # Characters in Private Use Area and above are ignored
-
-        section = codepoint >> 8   # Chop off the last two hex digits
-        position = codepoint % 256 # Last two hex digits
-
         try:
-            table = Cache[section]
+            retval.append(Cache_chars[char])
         except KeyError:
+            codepoint = ord(char)
+
+            if codepoint < 0x80:  # Basic ASCII
+                Cache_chars[char] = char
+                retval.append(char)
+                continue
+
+            if codepoint > 0xeffff:
+                continue  # Characters in Private Use Area and above are ignored
+
+            section = codepoint >> 8   # Chop off the last two hex digits
+            position = codepoint % 256  # Last two hex digits
+
             try:
-                mod = __import__('unidecode.x%03x'%(section), [], [], ['data'])
-            except ImportError:
-                Cache[section] = None
-                continue   # No match: ignore this character and carry on.
+                table = Cache_sections[section]
+            except KeyError:
+                try:
+                    mod = __import__('unidecode.x%03x' % (section), [], [], ['data'])
+                except ImportError:
+                    Cache_sections[section] = None
+                    continue   # No match: ignore this character and carry on.
 
-            Cache[section] = table = mod.data
+                Cache_sections[section] = table = mod.data
 
-        if table and len(table) > position:
-            retval.append( table[position] )
+            if table and len(table) > position:
+                Cache_chars[char] = table[position]
+                retval.append(table[position])
 
     return ''.join(retval)